Update dependencies
Fixes issue where using source maps with older versions of uglify generates
tons of warnings.

Using source maps with older versions of uglify generates many (~7000 for my
SPA) warnings that look like:

```
WARN: Couldn't figure out mapping for ?:62,2 → 1,1611 []
WARN: Couldn't figure out mapping for ?:62,15 → 1,1613 []
WARN: Couldn't figure out mapping for ?:63,4 → 1,1648 []
WARN: Couldn't figure out mapping for ?:68,9 → 1,1690 []
...
```

This is due to a [bug in the `source-map` library][bug] which is a dependency
of `uglify-js`.

Restricting `uglify-js` and `convert-source-map` to newer versions ensures that
incompatible versions of `source-map` are not installed.

[bug]: https://github.com/mishoo/UglifyJS2/issues/436